chore: refactor tokenlist to addressbook
